chore: pull transport construction into new file
